Update merge for MGA
Before we add adjoin functions, we have to make sure that the merge
functions do not reject merges based on the extra information each
selectional category is carrying along for adjunction.  I've done a
regression test, and this does not affect the correctness without
adjunction—however, with adjunction, these modifications will keep it
working.

License:       GPLv3+
Signed-off-by: Patrick M. Niedzielski <pmn25@cornell.edu>